+++ src/DataHelper.php
@@ -5,7 +5,7 @@ namespace Pronamic\WordPress\Pay\Gateways\IDealBasic;
 /**
  * Title: iDEAL Basic data helper class
  * Description:
- * Copyright: Copyright (c) 2005 - 2018
+ * Copyright: 2005-2019 Pronamic
  * Company: Pronamic
  *
  * @author  Remco Tolsma