+++ src/Item.php
@@ -7,7 +7,7 @@ use Pronamic\WordPress\Money\Money;
 /**
  * Title: iDEAL Basic item
  * Description:
- * Copyright: Copyright (c) 2005 - 2018
+ * Copyright: 2005-2019 Pronamic
  * Company: Pronamic
  *
  * @author  Remco Tolsma