+++ src/Notification.php
@@ -7,7 +7,7 @@ use DateTime;
 /**
  * Title: Notification
  * Description:
- * Copyright: Copyright (c) 2005 - 2018
+ * Copyright: 2005-2019 Pronamic
  * Company: Pronamic
  *
  * @author  Remco Tolsma